+++ shared/components/SideDrawer/SideDrawer.scss
@@ -64,7 +64,7 @@
     }
 
     .content-wrapper {
-      margin: 16px;
+      margin: 4px 16px;
       margin-left: calc(16px - var(--handleWidth));
     }
