+++ shared/contexts/YamlEditorContext/EditorActions.scss
@@ -1,0 +1,4 @@
+.circle-button {
+    border-radius: 50%;
+    background: rgba(193,193,193,0.2)!important;
+}
