+++ shared/contexts/YamlEditorContext/isValidYaml.js
@@ -1,0 +1,11 @@
+import jsyaml from 'js-yaml';
+
+export function isValidYaml(yaml) {
+  if (!yaml) return false;
+  try {
+    jsyaml.safeLoad(yaml);
+    return true;
+  } catch (e) {
+    return false;
+  }
+}
